# Reduced code generation in mat*vec and mat*mat multiply