--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -904,7 +904,7 @@ context("Matrix") do
     @fact b --> Mat([1,2,3,4]'')
 end
 context("Matrix Math") do
-	for i=1:4, j=1:4
+	for i=1:5, j=1:5
 		v = rand(j)
 		m = rand(i,j)
 		m2 = rand(i,j)