Heuristic thresholds to limit the size of generated expressions in
`mat*mat` and `mat*vec` loop unrolling.  The generated machine code still
isn't necessarily good here, but this improves compile times to
something acceptable in 0.4.

In julia 0.5 with `-O` the fully unrolled 4x4 multiplication does come out looking pretty nice with a lot of simd instructions being used.  Needs real benchmarking though.

@awbsmith @andyferris