@SimonDanisch 2016-07-14
Looking good!

Patched:
```
compile time 10x10 Float32: 
 0.998327 seconds (764.48 k allocations: 31.231 MB, 7.54% gc time)

Array{T,2} 2
BenchmarkTools.Trial:

  memory estimate:  128.00 bytes
  allocs estimate:  2
  minimum time:     70.00 ns (0.00% GC)
  median time:      73.00 ns (0.00% GC)
  mean time:        92.61 ns (16.37% GC)
  maximum time:     5.78 μs (97.92% GC)
Array{T,2} 4
BenchmarkTools.Trial:

  memory estimate:  176.00 bytes
  allocs estimate:  2
  minimum time:     320.00 ns (0.00% GC)
  median time:      326.00 ns (0.00% GC)
  mean time:        341.89 ns (2.69% GC)
  maximum time:     9.16 μs (94.10% GC)
Array{T,2} 10
BenchmarkTools.Trial:
  memory estimate:  576.00 bytes
  allocs estimate:  2
  minimum time:     655.00 ns (0.00% GC)
  median time:      669.00 ns (0.00% GC)
  mean time:        823.45 ns (2.57% GC)
  maximum time:     53.44 μs (0.00% GC)
FixedSizeArrays.Mat{Row,Column,T} 2
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     6.00 ns (0.00% GC)
  median time:      6.00 ns (0.00% GC)
  mean time:        6.44 ns (0.00% GC)
  maximum time:     60.00 ns (0.00% GC)
FixedSizeArrays.Mat{Row,Column,T} 4
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     10.00 ns (0.00% GC)
  median time:      11.00 ns (0.00% GC)
  mean time:        11.61 ns (0.00% GC)
  maximum time:     166.00 ns (0.00% GC)
FixedSizeArrays.Mat{Row,Column,T} 10
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     2.73 μs (0.00% GC)
  median time:      2.73 μs (0.00% GC)
  mean time:        2.86 μs (0.00% GC)
  maximum time:     10.12 μs (0.00% GC)
```

unpatched:
```
compile time 10x10 Float32:  3.945150 seconds (3.26 M allocations: 92.470 MB, 2.52% gc time)
FixedSizeArrays.Mat{Row,Column,T} 2
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     6.00 ns (0.00% GC)
  median time:      7.00 ns (0.00% GC)
  mean time:        7.25 ns (0.00% GC)
  maximum time:     34.00 ns (0.00% GC)
FixedSizeArrays.Mat{Row,Column,T} 4
BenchmarkTools.Trial:

  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     10.00 ns (0.00% GC)
  median time:      11.00 ns (0.00% GC)
  mean time:        11.84 ns (0.00% GC)
  maximum time:     77.00 ns (0.00% GC)
FixedSizeArrays.Mat{Row,Column,T} 10
BenchmarkTools.Trial:

  memory estimate:  15.63 kb
  allocs estimate:  1000
  minimum time:     15.17 μs (0.00% GC)
  median time:      16.07 μs (0.00% GC)
  mean time:        20.32 μs (6.31% GC)
  maximum time:     2.74 ms (97.52% GC)
```

benchmark code:
```Julia
using FixedSizeArrays, BenchmarkTools
function rand_matrix{X<:Matrix}(A::Type{X}, N, T)
    rand(T, N, N)
end
function rand_matrix{X<:FixedSizeArrays.Mat}(A::Type{X}, N, T)
    rand(A{N, N, T})
end
constructors = [
    ("Array", Matrix),
    ("FSA", FixedSizeArrays.Mat),
]
for (name, c) in constructors, N in (2, 4, 10)
    T = Float32
      a = rand_matrix(c, N, T)
      b = rand_matrix(c, N, T)
      println(c, " ", N)
      println(@benchmark $a * $b)
end
```