@andyferris 2016-07-14
I just re-ran *StaticArrays* benchmarks after checking out the `matmul-performance` branch. Initial indications are that it's still 10x slower for 4x4 matrix-matrix multiplications (unless I've done something stupid).

```
=====================================
    Benchmarks for 4×4 matrices
=====================================
StaticArrays compilation time (×3):  0.833219 seconds (414.44 k allocations: 17.265 MB, 7.44% gc time)
FixedSizeArrays compilation time:    0.815958 seconds (662.31 k allocations: 28.506 MB, 0.73% gc time)

Matrix multiplication and accumulation
--------------------------------------
Array             ->  6.019106 seconds (62.50 M allocations: 6.985 GB, 4.20% gc time)
Array (mutating)  ->  4.310547 seconds (7 allocations: 608 bytes)
SArray            ->  0.047942 seconds (5 allocations: 304 bytes)
MArray            ->  0.725796 seconds (15.63 M allocations: 2.095 GB, 9.79% gc time)
MArray (mutating) ->  0.532509 seconds (6 allocations: 448 bytes)
Mat               ->  0.476407 seconds (5 allocations: 304 bytes)

Matrix addition and accumulation
--------------------------------
Array             ->  1.301290 seconds (25.00 M allocations: 2.794 GB, 8.12% gc time)
Array (mutating)  -> 18.450668 seconds (112.50 M allocations: 2.049 GB, 1.81% gc time)
SArray            ->  0.037575 seconds (5 allocations: 304 bytes)
MArray            ->  0.385095 seconds (12.50 M allocations: 1.676 GB, 13.51% gc time)
MArray (mutating) ->  0.125285 seconds (5 allocations: 304 bytes)
Mat               ->  0.042882 seconds (5 allocations: 304 bytes)
```