@c42f 2016-07-14
@andyferris yes, this patch was more to address the compile time woes (more than 30s to compile a 12x12 matmul in 0.4?!) rather than the runtime, but it's worth getting to the bottom of both.

Given the assembly I saw with `julia-0.5 -O` and `Mat{4,4,Float32}`, I'd be rather surprised if a 10x speedup is actually possible, since the codegen looked pretty good.  I'm suspicious something is being optimized away here.

OTOH, `Mat{10,10,Float32}` is clearly still quite bad according to Simon's comparison with Array above...

Complicating this whole thing is the fact that julia's Tuple-related codegen behaves completely differently in 0.4 and 0.5.