@andyferris 2016-07-14
> this patch was more to address the compile time woes

Oh right! Compile time seems similar on 0.5 (that's qualitative, not quantitative).

>  I'd be rather surprised if a 10x speedup is actually possible, since the codegen looked pretty good. I'm suspicious something is being optimized away here.

That's what I was referring to be doing something stupid, though I did try to investigate this. Both functions are accumulators, so I'm figuring *something* has to occur in each loop. Any more eyes on [my benchmark](https://github.com/andyferris/StaticArrays.jl/blob/master/perf/benchmark.jl) would be greatly appreciated. :)