@c42f 2016-07-14
Yes, right.  I suspect the compiler is pulling some code out of your accumulation loop on line 30 by proving it's loop invariant.  If this is possible for StaticArrays but not FixedSizeArrays, it's yet another sign that the non-nested approach of StaticArrays is the right thing, since the compiler is having an easier time reasoning about it.

I checked the generated assembly for multiplication of 4x4 matrices on my local machine and it's *exactly* the same for both packages, bar a line number.