@c42f 2016-07-14
An easy way to defeat the loop invariance optimization is probably just to use something like:

```julia
@noinline function innerloop(As, Bs)
    C = As[1]*Bs[1]
    for i=2:length(As)
        C += As[i]*Bs[i]
    end
    C
end
```

Then call this function with a pair of (short!) Vectors of `Mat`, possibly multiple times.  This might be a bit more representative

Probably better, check out *BenchmarkTools.jl* and just use the `@benchmark` macro since jrevels has done a bunch of hard work there to make benchmarking of base more reliable.