@SimonDanisch 2016-07-14
Even with this patch it's still completely unrolled, we're just saving a bit in compile time by inlining less.
And unrolling becomes slow at some point... To make matmul fast for large matrices we'll need loops and mutable fixed arrays.

>it's yet another sign that the non-nested approach of StaticArrays is the right thing

I wouldn't mind that... I was a bit sad that I needed to switch to nested tuples... I'm actually not sure anymore what the concrete reasons where, but I remembered doing a few tests and the nested approach was strictly better, besides implementation problems ;) But this might have changed! :)